Soft-deprecate params
